NTUSE-60 First Run data population for profile creation and update database
If we have existing data in the auth user object itself we will pull it out for usage. This includes the Display Name and Profile Picture for Google Logins. This way the user does not have so much trouble on this
We also added code to add the user object to the database and if successful, send the user directly over to the main activity
We also renamed the checker activity to updating user activity. This activity will be used to process all of the on startup tasks we have to execute if any