NTUSE-99 Added ability to click on drawer header to view your profile
The view profile page will be the same as the edit page except that we cannot edit any data. At least until we click on the "edit" icon

While the edit icon is not implemented yet, what I plan to do is to change the FAB icon to an edit icon and when clicked, allow editing of the profile. Thereafter the same FAB icon will switch over to a save icon