NTUSE-70 Updated method to favourite a gym
Instead of storing it into the user object we can instead just store straight into a new collection. This way we are able to "shunbian" calculate the favourite count for each gym as well
Also by doing it this way we also allow us to display the favourites rating easier too so winwin! :D